@@ -4,32 +4,31 @@ using Microsoft.Extensions.Hosting;
 using Formality.Api.Extensions;
 using Formality.App.Infrastructure;
 
-namespace Formality.Api
+namespace Formality.Api;
+
+public class Program
 {
-    public class Program
+    public static void Main(string[] args)
     {
-        public static void Main(string[] args)
+        var host = CreateHostBuilder(args).Build();
+        var env = host.Services.GetRequiredService<IWebHostEnvironment>();
+
+        if (env.IsDevelopment())
         {
-            var host = CreateHostBuilder(args).Build();
-            var env = host.Services.GetRequiredService<IWebHostEnvironment>();
-
-            if (env.IsDevelopment())
+            host.MigrateDbContext<AppDbContext>((_, services) =>
             {
-                host.MigrateDbContext<AppDbContext>((_, services) =>
-                {
-                    var seed = services.GetRequiredService<AppDbContextSeed>();
-                    seed.SeedAsync().GetAwaiter().GetResult();
-                });
-            }
-
-            host.Run();
+                var seed = services.GetRequiredService<AppDbContextSeed>();
+                seed.SeedAsync().GetAwaiter().GetResult();
+            });
         }
 
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder =>
-                {
-                    webBuilder.UseStartup<Startup>();
-                });
+        host.Run();
     }
+
+    public static IHostBuilder CreateHostBuilder(string[] args) =>
+        Host.CreateDefaultBuilder(args)
+            .ConfigureWebHostDefaults(webBuilder =>
+            {
+                webBuilder.UseStartup<Startup>();
+            });
 }
