@@ -1,9 +1,8 @@
-namespace Formality.App.Common.Dto
-{
-    public class OrderDto
-    {
-        public string Name { get; set; } = string.Empty;
+namespace Formality.App.Common.Dto;
 
-        public bool Desc { get; set; }
-    }
+public class OrderDto
+{
+    public string Name { get; set; } = string.Empty;
+
+    public bool Desc { get; set; }
 }
