@@ -1,21 +1,20 @@
 using System;
 
-namespace Formality.App.Common.Exceptions
+namespace Formality.App.Common.Exceptions;
+
+public class DomainException : Exception
 {
-    public class DomainException : Exception
+    public DomainException()
     {
-        public DomainException()
-        {
-        }
+    }
 
-        public DomainException(string message)
-            : base(message)
-        {
-        }
+    public DomainException(string message)
+        : base(message)
+    {
+    }
 
-        public DomainException(string message, Exception innerException)
-            : base(message, innerException)
-        {
-        }
+    public DomainException(string message, Exception innerException)
+        : base(message, innerException)
+    {
     }
 }
