@@ -1,7 +1,6 @@
-namespace Formality.App.Common.Models
+namespace Formality.App.Common.Models;
+
+public abstract class Entity
 {
-    public abstract class Entity
-    {
-        public int Id { get; set; }
-    }
+    public int Id { get; set; }
 }
