@@ -7,49 +7,48 @@ using Formality.App.Infrastructure;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 
-namespace Formality.App.Forms.Commands
+namespace Formality.App.Forms.Commands;
+
+public class CreateFormCommand : IRequest<int>
 {
-    public class CreateFormCommand : IRequest<int>
+    public string Name { get; set; } = string.Empty;
+}
+
+public sealed class CreateFormCommandHandler : IRequestHandler<CreateFormCommand, int>
+{
+    private readonly AppDbContext _context;
+
+    public CreateFormCommandHandler(AppDbContext context)
     {
-        public string Name { get; set; } = string.Empty;
+        _context = context;
     }
 
-    public sealed class CreateFormCommandHandler : IRequestHandler<CreateFormCommand, int>
+    public async Task<int> Handle(
+        CreateFormCommand request,
+        CancellationToken cancellationToken)
     {
-        private readonly AppDbContext _context;
-
-        public CreateFormCommandHandler(AppDbContext context)
+        if (await _context.Forms.AnyAsync(x => x.Name == request.Name, cancellationToken))
         {
-            _context = context;
+            throw new DomainException($@"Form ""{request.Name}"" already exists.");
         }
 
-        public async Task<int> Handle(
-            CreateFormCommand request,
-            CancellationToken cancellationToken)
+        var form = new Form
         {
-            if (await _context.Forms.AnyAsync(x => x.Name == request.Name, cancellationToken))
-            {
-                throw new DomainException($@"Form ""{request.Name}"" already exists.");
-            }
+            Name = request.Name,
+        };
 
-            var form = new Form
-            {
-                Name = request.Name,
-            };
+        _context.Add(form);
 
-            _context.Add(form);
+        await _context.SaveChangesAsync(CancellationToken.None);
 
-            await _context.SaveChangesAsync(CancellationToken.None);
-
-            return form.Id;
-        }
+        return form.Id;
     }
+}
 
-    public sealed class CreateFormCommandValidator : AbstractValidator<EditFormCommand>
+public sealed class CreateFormCommandValidator : AbstractValidator<EditFormCommand>
+{
+    public CreateFormCommandValidator()
     {
-        public CreateFormCommandValidator()
-        {
-            RuleFor(x => x.Name).NotEmpty().MaximumLength(256);
-        }
+        RuleFor(x => x.Name).NotEmpty().MaximumLength(256);
     }
 }
