@@ -3,16 +3,15 @@ using System.Linq;
 using Formality.App.Common.Dto;
 using Formality.App.Forms.Models;
 
-namespace Formality.App.Forms.Dto
+namespace Formality.App.Forms.Dto;
+
+public class FormDto : NamedEntityDto
 {
-    public class FormDto : NamedEntityDto
-    {
-        public FormState StateId { get; set; }
+    public FormState StateId { get; set; }
 
-        public IEnumerable<FormFieldDto> Fields { get; set; } = Enumerable.Empty<FormFieldDto>();
-    }
+    public IEnumerable<FormFieldDto> Fields { get; set; } = Enumerable.Empty<FormFieldDto>();
+}
 
-    public class FormListDto : NamedEntityDto
-    {
-    }
+public class FormListDto : NamedEntityDto
+{
 }
