@@ -1,9 +1,8 @@
-namespace Formality.App.Forms.Dto
-{
-    public class FormFieldRuleDto : RuleDto
-    {
-        public int Id { get; set; }
+namespace Formality.App.Forms.Dto;
 
-        public int FieldId { get; set; }
-    }
+public class FormFieldRuleDto : RuleDto
+{
+    public int Id { get; set; }
+
+    public int FieldId { get; set; }
 }
