@@ -1,9 +1,8 @@
-namespace Formality.App.Forms.Dto
-{
-    public sealed class FormFieldValueDto
-    {
-        public int? Id { get; set; }
+namespace Formality.App.Forms.Dto;
 
-        public string Value { get; set; } = string.Empty;
-    }
+public sealed class FormFieldValueDto
+{
+    public int? Id { get; set; }
+
+    public string Value { get; set; } = string.Empty;
 }
