@@ -1,12 +1,11 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Formality.App.Forms.Dto
-{
-    public sealed class FormFieldValuesDto
-    {
-        public int FormId { get; set; }
+namespace Formality.App.Forms.Dto;
 
-        public IEnumerable<FieldValueDto> Values { get; set; } = Enumerable.Empty<FieldValueDto>();
-    }
+public sealed class FormFieldValuesDto
+{
+    public int FormId { get; set; }
+
+    public IEnumerable<FieldValueDto> Values { get; set; } = Enumerable.Empty<FieldValueDto>();
 }
