@@ -1,11 +1,10 @@
 using Formality.App.Forms.Models;
 
-namespace Formality.App.Forms.Dto
-{
-    public class RuleDto
-    {
-        public FieldRule Type { get; set; }
+namespace Formality.App.Forms.Dto;
 
-        public string? Data { get; set; }
-    }
+public class RuleDto
+{
+    public FieldRule Type { get; set; }
+
+    public string? Data { get; set; }
 }
