@@ -1,16 +1,15 @@
-namespace Formality.App.Forms.Models
+namespace Formality.App.Forms.Models;
+
+public enum FieldRule
 {
-    public enum FieldRule
-    {
-        Unknown,
-        Required,
-        Length,
-    }
+    Unknown,
+    Required,
+    Length,
+}
 
-    public class LengthData
-    {
-        public int MinLength { get; set; }
+public class LengthData
+{
+    public int MinLength { get; set; }
 
-        public int MaxLength { get; set; }
-    }
+    public int MaxLength { get; set; }
 }
