@@ -1,14 +1,13 @@
-using Formality.App.Common.Models;
 using System.Collections.Generic;
+using Formality.App.Common.Models;
 
-namespace Formality.App.Forms.Models
+namespace Formality.App.Forms.Models;
+
+public class Form : Entity
 {
-    public class Form : Entity
-    {
-        public string Name { get; set; } = string.Empty;
+    public string Name { get; set; } = string.Empty;
 
-        public FormState StateId { get; set; } = FormState.New;
+    public FormState StateId { get; set; } = FormState.New;
 
-        public ICollection<FormField> Fields { get; set; } = new List<FormField>();
-    }
+    public ICollection<FormField> Fields { get; set; } = new List<FormField>();
 }
