@@ -1,13 +1,12 @@
 using Formality.App.Common.Models;
 
-namespace Formality.App.Forms.Models
+namespace Formality.App.Forms.Models;
+
+public class FormFieldRule : Entity
 {
-    public class FormFieldRule : Entity
-    {
-        public FieldRule Type { get; set; }
+    public FieldRule Type { get; set; }
 
-        public FormField Field { get; set; } = default!;
+    public FormField Field { get; set; } = default!;
 
-        public string? Data { get; set; }
-    }
+    public string? Data { get; set; }
 }
