@@ -4,105 +4,104 @@ using System.Threading.Tasks;
 using Formality.App.Forms.Models;
 using Microsoft.EntityFrameworkCore;
 
-namespace Formality.App.Infrastructure
+namespace Formality.App.Infrastructure;
+
+public class AppDbContextSeed
 {
-    public class AppDbContextSeed
+    private readonly AppDbContext _context;
+
+    public AppDbContextSeed(AppDbContext context)
     {
-        private readonly AppDbContext _context;
+        _context = context;
+    }
 
-        public AppDbContextSeed(AppDbContext context)
+    public virtual async Task SeedAsync()
+    {
+        if (!await _context.Forms.AnyAsync())
         {
-            _context = context;
+            await AddDefaultForm();
         }
+    }
 
-        public virtual async Task SeedAsync()
+    private async Task AddDefaultForm()
+    {
+        var form = new Form
         {
-            if (!await _context.Forms.AnyAsync())
+            Name = "Default Form",
+            StateId = FormState.Actual,
+            Fields = new List<FormField>
             {
-                await AddDefaultForm();
-            }
-        }
-
-        private async Task AddDefaultForm()
-        {
-            var form = new Form
-            {
-                Name = "Default Form",
-                StateId = FormState.Actual,
-                Fields = new List<FormField>
+                new FormField
                 {
-                    new FormField
+                    Name = "name",
+                    Label = "Name",
+                    Placeholder = "Please enter your name",
+                    Type = FieldType.SingleLineText,
+                    Rules = new List<FormFieldRule>
                     {
-                        Name = "name",
-                        Label = "Name",
-                        Placeholder = "Please enter your name",
-                        Type = FieldType.SingleLineText,
-                        Rules = new List<FormFieldRule>
+                        new FormFieldRule
                         {
-                            new FormFieldRule
-                            {
-                                Type = FieldRule.Required
-                            },
-                        },
-                    },
-                    new FormField
-                    {
-                        Name = "about",
-                        Label = "About",
-                        Placeholder = "Please tell us something about yourself",
-                        Type = FieldType.MultiLineText,
-                        Rules = new List<FormFieldRule>
-                        {
-                            new FormFieldRule
-                            {
-                                Type = FieldRule.Length,
-                                Data = JsonSerializer.Serialize(
-                                    new LengthData
-                                    {
-                                        MaxLength = 200,
-                                        MinLength = 2,
-                                    },
-                                    new JsonSerializerOptions
-                                    {
-                                        PropertyNamingPolicy = JsonNamingPolicy.CamelCase
-                                    }),
-                            },
-                        },
-                    },
-                    new FormField
-                    {
-                        Name = "quantity",
-                        Label = "Quantity",
-                        Type = FieldType.Dropdown,
-                        Rules = new List<FormFieldRule>
-                        {
-                            new FormFieldRule
-                            {
-                                Type = FieldRule.Required
-                            },
-                        },
-                        Values = new List<FormFieldValue>
-                        {
-                            new FormFieldValue
-                            {
-                                Value = "One",
-                            },
-                            new FormFieldValue
-                            {
-                                Value = "Hundred",
-                            },
-                            new FormFieldValue
-                            {
-                                Value = "Thousand",
-                            },
+                            Type = FieldRule.Required
                         },
                     },
                 },
-            };
+                new FormField
+                {
+                    Name = "about",
+                    Label = "About",
+                    Placeholder = "Please tell us something about yourself",
+                    Type = FieldType.MultiLineText,
+                    Rules = new List<FormFieldRule>
+                    {
+                        new FormFieldRule
+                        {
+                            Type = FieldRule.Length,
+                            Data = JsonSerializer.Serialize(
+                                new LengthData
+                                {
+                                    MaxLength = 200,
+                                    MinLength = 2,
+                                },
+                                new JsonSerializerOptions
+                                {
+                                    PropertyNamingPolicy = JsonNamingPolicy.CamelCase
+                                }),
+                        },
+                    },
+                },
+                new FormField
+                {
+                    Name = "quantity",
+                    Label = "Quantity",
+                    Type = FieldType.Dropdown,
+                    Rules = new List<FormFieldRule>
+                    {
+                        new FormFieldRule
+                        {
+                            Type = FieldRule.Required
+                        },
+                    },
+                    Values = new List<FormFieldValue>
+                    {
+                        new FormFieldValue
+                        {
+                            Value = "One",
+                        },
+                        new FormFieldValue
+                        {
+                            Value = "Hundred",
+                        },
+                        new FormFieldValue
+                        {
+                            Value = "Thousand",
+                        },
+                    },
+                },
+            },
+        };
 
-            _context.Add(form);
+        _context.Add(form);
 
-            await _context.SaveChangesAsync();
-        }
+        await _context.SaveChangesAsync();
     }
 }
