@@ -4,20 +4,19 @@ using Formality.App.Submissions.Commands;
 using Formality.App.Submissions.Dto;
 using Formality.App.Submissions.Models;
 
-namespace Formality.App.Infrastructure.AutoMapper
+namespace Formality.App.Infrastructure.AutoMapper;
+
+public class SubmissionsProfile : Profile
 {
-    public class SubmissionsProfile : Profile
+    public SubmissionsProfile()
     {
-        public SubmissionsProfile()
-        {
-            CreateMap<Submission, SubmissionDto>()
-                .ForMember(x => x.Values, x => x.ExplicitExpansion());
-            CreateMap<Submission, SubmissionListDto>();
+        CreateMap<Submission, SubmissionDto>()
+            .ForMember(x => x.Values, x => x.ExplicitExpansion());
+        CreateMap<Submission, SubmissionListDto>();
 
-            CreateMap<SubmissionValue, SubmissionValueDto>();
-            CreateMap<SubmissionValueDto, FieldValueDto>();
+        CreateMap<SubmissionValue, SubmissionValueDto>();
+        CreateMap<SubmissionValueDto, FieldValueDto>();
 
-            CreateMap<AddSubmissionCommand, FormFieldValuesDto>();
-        }
+        CreateMap<AddSubmissionCommand, FormFieldValuesDto>();
     }
 }
