@@ -3,35 +3,39 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using FluentValidation;
+using FluentValidation.Results;
 using MediatR;
 
-namespace Formality.App.Infrastructure.MediatR
+namespace Formality.App.Infrastructure.MediatR;
+
+public class ValidationBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
+    where TRequest : IRequest<TResponse>
 {
-    public class ValidationBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
-        where TRequest : IRequest<TResponse>
+    private readonly IEnumerable<IValidator<TRequest>> _validators;
+
+    public ValidationBehavior(IEnumerable<IValidator<TRequest>> validators)
     {
-        private readonly IEnumerable<IValidator<TRequest>> _validators;
+        _validators = validators;
+    }
 
-        public ValidationBehavior(IEnumerable<IValidator<TRequest>> validators)
+    public async Task<TResponse> Handle(
+        TRequest request,
+        CancellationToken cancellationToken,
+        RequestHandlerDelegate<TResponse> next)
+    {
+        var errors = new List<ValidationFailure>();
+
+        foreach (var validator in _validators)
         {
-            _validators = validators;
+            var result = await validator.ValidateAsync(request, cancellationToken);
+            errors.AddRange(result.Errors);
         }
 
-        public async Task<TResponse> Handle(
-            TRequest request,
-            CancellationToken cancellationToken,
-            RequestHandlerDelegate<TResponse> next)
+        if (errors.Any())
         {
-            var tasks = _validators.Select(x => x.ValidateAsync(request, cancellationToken));
-            var results = await Task.WhenAll(tasks);
-            var errors = results.SelectMany(x => x.Errors);
-
-            if (errors.Any())
-            {
-                throw new ValidationException(errors);
-            }
-
-            return await next();
+            throw new ValidationException(errors);
         }
+
+        return await next();
     }
 }
