@@ -1,9 +1,5 @@
-using System;
-using System.Collections.Generic;
-using System.IO;
 using System.Linq;
-using System.Net.Http;
-using System.Text;
+using System.Net.Http.Json;
 using System.Text.Json;
 using System.Threading.Tasks;
 using FluentAssertions;
@@ -13,49 +9,48 @@ using Formality.App.Forms.Queries;
 using Formality.Tests.Fixtures;
 using Xunit;
 
-namespace Formality.Tests.FunctionalTests
+namespace Formality.Tests.FunctionalTests;
+
+[Collection(nameof(ControllerTestBase))]
+public class FormsControllerTests : ControllerTestBase
 {
-    [Collection(nameof(ControllerTestBase))]
-    public class FormsControllerTests : ControllerTestBase
+    private readonly string _requestUri = "/api/forms";
+
+    public FormsControllerTests(WebApplicationFixture factory) : base(factory)
     {
-        private readonly string _requestUri = "/api/forms";
+    }
 
-        public FormsControllerTests(WebApplicationFixture factory) : base(factory)
+    [Fact]
+    public async Task SearchForms_NoExceptions()
+    {
+        var keyword = "Default";
+        var queryString = GetQueryString(keyword);
+
+        var response = await Client.GetAsync(_requestUri + $"?{queryString}");
+
+        response.EnsureSuccessStatusCode();
+
+        var forms = await DeserializeResponse<FormListDto[]>(response);
+        var form = forms.FirstOrDefault(x => x.Name.Contains(keyword));
+
+        form.Should().NotBeNull();
+    }
+
+    private string GetQueryString(string keyword)
+    {
+        var query = new SearchFormQuery
         {
-        }
-
-        [Fact]
-        public async Task SearchForms_NoExceptions()
-        {
-            var keyword = "Default";
-            var queryString = GetQueryString(keyword);
-
-            var response = await Client.GetAsync(_requestUri + $"?{queryString}");
-
-            response.EnsureSuccessStatusCode();
-
-            var forms = await DeserializeResponse<FormListDto[]>(response);
-            var form = forms.FirstOrDefault(x => x.Name.Contains(keyword));
-
-            form.Should().NotBeNull();
-        }
-
-        private string GetQueryString(string keyword)
-        {
-            var query = new SearchFormQuery
+            Keyword = keyword,
+            MaxResults = 10,
+            OrderBy = new[]
             {
-                Keyword = keyword,
-                MaxResults = 10,
-                OrderBy = new[]
-                {
-                    new OrderDto { Name = "Id", Desc = true }
-                }
-            };
+                new OrderDto { Name = "Id", Desc = true }
+            }
+        };
 
-            return JsonSerializer.Serialize(query, new JsonSerializerOptions
-            {
-                PropertyNamingPolicy = JsonNamingPolicy.CamelCase
-            });
-        }
+        return JsonSerializer.Serialize(query, new JsonSerializerOptions
+        {
+            PropertyNamingPolicy = JsonNamingPolicy.CamelCase
+        });
     }
 }
